añadiendo componentes por separado para el nav y pruebas de diseño para la pagina
