+++ src/components/landing/Hero.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import '../../styles/Layout.css';
+
+function Hero() {
+  return (
+    <>
+      <section className="hero">
+            <h1>Hello World</h1>
+            <ul>
+                  <li><a href="">More</a></li>
+                  <li><a href="">Join</a></li>
+            </ul>
+      </section>
+    </>
+  );
+}
+
+export default Hero;
