--- src/components/landing/Link.jsx
@@ -1,12 +1,0 @@
-import React from 'react'
-
-function Link() {
-      
-  return (
-    <div>
-      
-    </div>
-  )
-}
-
-export default Link
