+++ src/components/nav/logo.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function logo() {
+  return (
+    <>
+      <a href="#"><h3>Logo</h3></a>
+    </>
+  )
+}
+
+export default logo;
