+++ src/styles/Layout.css
@@ -1,15 +1,81 @@
+/*import fonts and others*/
+
+/*<<------------------------------------->>*/
+
+/*generals*/
+*{
+      margin: 0;
+      padding: 0;
+      box-sizing: border-box;
+}
+h1,h2,h3,h4,h5,h6{
+      margin: 0;
+      color: fff;
+      font-weight: bolder;
+}
+
+p{
+      color: #fff;
+      text-align: center;
+}
+
+li{list-style: none;}
+
+a{
+      text-decoration: none;
+      color: #fff;
+}
+
 body{
+      width: 100vw;
       height: 100vh;
-      background-color: #202020;
       color: #fff;
       display: flex;
       justify-content: center;
+      align-items: flex-start;
+      background-color: #282020;
 }
 
-.web-container{
+/*Component/nav/NavBar.jsx*/
+header{
       display: flex;
-      max-width: 1600px;
+      justify-content: space-around;
+      width: 100%;
+      max-height: 50px;
+      margin: 5px 50px;
+      align-items: center;
+}
+
+/*layouts/Layout.astro*/
+.web-container{ 
+      display: flex;
+      max-width: 1280px;
       width: 100%;
       flex-wrap: wrap;
+      position: relative;
+}
 
+main{
+      width: 100%;
+      height: 100vh;
+}
+
+/*component/nav/Link.jsx*/
+.nav-bar{ 
+      display: flex;
+      width: 30vw;
+      align-items: center;
+      justify-content: space-between;
+}
+
+/*component/landing/hero*/
+.hero{
+      background-image: url(../assets/backGroundBody.png);
+      background-size: cover;
+      background-repeat: no-repeat;
+      width: 100%;
+      height: 400px;
+      display: flex;
+      flex-wrap: wrap;
+      border-radius: 20px;
 }
